delete, update methods created

+++ acme-flight-system/DataBaseManager.cs
@@ -26,6 +26,8 @@ namespace acme_flight_system
 
         public async Task AdicionaVoo(VooModel novo_voo,System.Action populate_grid_view_callback)
         {
+            var flight_exists = await IsFlightAlreadyRegistered(novo_voo);
+            if (flight_exists) return;
             string query = "insert into TB_VOO values " +
                         "(@ID_VOO,@Data_voo,@Custo,@Distancia,@Captura,@Nivel_dor)";
             using (var database_connection = GetDbConnection())
@@ -35,13 +37,25 @@ namespace acme_flight_system
             }
         }
 
-        public async Task DeleteVoo(VooModel novo_voo)
+        public async Task<bool> IsFlightAlreadyRegistered(VooModel voo)
         {
-            string query = "DELETE FROM TB_VOO" +
-                           "WHERE ID_VOO=@ID_VOO";
+            string query = "SELECT * FROM TB_VOO WHERE ID_VOO = @ID_VOO";
             using (var database_connection = GetDbConnection())
             {
-                await Task.Run(() => database_connection.Execute(query, novo_voo));
+                var filght_rows = await Task.Run(() => database_connection.Execute(query,voo));
+                return filght_rows > 0;
+            }
+        }
+
+        public async Task DeleteVoo(VooModel voo)
+        {
+            var flight_exists = await IsFlightAlreadyRegistered(voo);
+            if (!flight_exists) return;
+            string query = "DELETE FROM TB_VOO" +
+                           "WHERE ID_VOO= @ID_VOO";
+            using (var database_connection = GetDbConnection())
+            {
+                await Task.Run(() => database_connection.Execute(query, voo));
             }
         }
